+++ src/main/java/com/damageddream/medicalclinic/service/DoctorService.java
@@ -11,6 +11,5 @@ public interface DoctorService {
     DoctorDTO save(NewDoctorDTO newDoctorDTO);
     DoctorDTO addFacilityToDoctor(Long doctorId, GetIdCommand entityId);
 
-    AppointmentDTO addAppointment(Long doctorId, Appointment appointment);
-    List<AppointmentDTO> getFreeAppointmentsByDoctor(Long doctorId);
+
 }
