+++ src/main/java/com/damageddream/medicalclinic/service/PatientServiceImpl.java
@@ -25,8 +25,7 @@ public class PatientServiceImpl implements PatientService {
     private final DataValidator dataValidator;
     private final PatientMapper patientMapper;
     private final PatientRepository patientRepository;
-    private final AppointmentMapper appointmentMapper;
-    private final AppointmentRepository appointmentRepository;
+
 
     @Override
     @Transactional
@@ -85,32 +84,5 @@ public class PatientServiceImpl implements PatientService {
         return patientMapper.toDTO(toEdit);
     }
 
-    @Override
-    @Transactional
-    public AppointmentDTO makeAnAppointment(Long patientId, GetIdCommand appointmentIdCommand) {
-        Long appointmentId = appointmentIdCommand.getEntityId();
-        Patient patient = patientRepository.findById(patientId)
-                .orElseThrow(()->new PatientNotFoundException("Patient not found"));
-        Appointment appointment = appointmentRepository.findById(appointmentId).orElseThrow(
-                ()-> new AppointmentNotFoundException("There is no such appointment")
-        );
-        if(appointment.getPatient() != null){
-            throw  new AppointmentNotFoundException("That appointment is no longer free");
-        }
-        appointment.setPatient(patient);
-        return appointmentMapper.toDTO(appointment);
-    }
 
-    @Override
-    public List<AppointmentDTO> getPatientsAppointments(Long id) {
-        Patient patient = patientRepository.findById(id)
-                .orElseThrow(()->new PatientNotFoundException("Patient not found"));
-
-        List<Appointment> appointments = appointmentRepository.findAppointmentsByPatient(id);
-        if(appointments.isEmpty()){
-            throw new AppointmentNotFoundException("Patient don't have any appointments");
-        }
-
-        return appointments.stream().map(appointmentMapper::toDTO).toList();
-    }
 }
