+++ src/test/java/com/damageddream/medicalclinic/dto/mapper/AppointmentMapperTest.java
@@ -1,0 +1,4 @@
+package com.damageddream.medicalclinic.dto.mapper;
+
+public class AppointmentMapperTest {
+}
