+++ src/test/java/com/damageddream/medicalclinic/service/DoctorServiceTest.java
@@ -2,7 +2,6 @@ package com.damageddream.medicalclinic.service;
 
 import com.damageddream.medicalclinic.dto.GetIdCommand;
 import com.damageddream.medicalclinic.dto.NewDoctorDTO;
-import com.damageddream.medicalclinic.dto.mapper.AppointmentMapper;
 import com.damageddream.medicalclinic.dto.mapper.DoctorMapper;
 import com.damageddream.medicalclinic.dto.mapper.FacilityMapper;
 import com.damageddream.medicalclinic.entity.Doctor;
@@ -10,11 +9,9 @@ import com.damageddream.medicalclinic.entity.Facility;
 import com.damageddream.medicalclinic.exception.DoctorAlreadyExistsException;
 import com.damageddream.medicalclinic.exception.DoctorNotFoundException;
 import com.damageddream.medicalclinic.exception.FacilityNotFoundException;
-import com.damageddream.medicalclinic.repository.AppointmentRepository;
 import com.damageddream.medicalclinic.repository.DoctorRepository;
 import com.damageddream.medicalclinic.repository.FacilityRepository;
 import com.damageddream.medicalclinic.util.TestDataFactory;
-import com.damageddream.medicalclinic.validation.DataValidator;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.mapstruct.factory.Mappers;
@@ -31,9 +28,7 @@ public class DoctorServiceTest {
     private DoctorMapper doctorMapper;
     private FacilityRepository facilityRepository;
     private DoctorRepository doctorRepository;
-    private AppointmentRepository appointmentRepository;
-    private DataValidator dataValidator;
-    private AppointmentMapper appointmentMapper;
+
 
     private DoctorService doctorService;
 
@@ -45,8 +40,8 @@ public class DoctorServiceTest {
         this.doctorRepository = Mockito.mock(DoctorRepository.class);
 
 
-        this.doctorService = new DoctorServiceImpl(doctorRepository, facilityRepository, appointmentRepository,
-                doctorMapper, facilityMapper, dataValidator, appointmentMapper);
+        this.doctorService = new DoctorServiceImpl(doctorRepository, facilityRepository,
+                doctorMapper, facilityMapper);
     }
 
     @Test
